+++ ContentPatcher/Framework/Tokens/IToken.cs
@@ -19,6 +19,9 @@ namespace ContentPatcher.Framework.Tokens
         /// <summary>Whether this token is only valid with input arguments.</summary>
         bool RequiresInput { get; }
 
+        /// <summary>Whether to allow using this token in any value context (e.g. as a number or boolean) without validating ahead of time.</summary>
+        bool BypassesContextValidation { get; }
+
 
         /*********
         ** Public methods
@@ -56,6 +59,10 @@ namespace ContentPatcher.Framework.Tokens
         /// <param name="max">The maximum value this token may return.</param>
         /// <exception cref="InvalidOperationException">The input doesn't match this value provider.</exception>
         bool HasBoundedRangeValues(IInputArguments input, out int min, out int max);
+
+        /// <summary>Normalize a raw value so it can be compared with the token values.</summary>
+        /// <param name="value">The raw value.</param>
+        string NormalizeValue(string value);
 
         /// <summary>Get the current token values.</summary>
         /// <param name="input">The input arguments.</param>