+++ ContentPatcher/Framework/Tokens/ValueProviders/BaseValueProvider.cs
@@ -48,6 +48,9 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
         /// <inheritdoc />
         public bool RequiresPositionalInput { get; private set; }
 
+        /// <inheritdoc />
+        public bool BypassesContextValidation { get; protected set; } = false;
+
 
         /*********
         ** Public methods
@@ -205,6 +208,12 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
         {
             this.AssertInput(input);
             yield break;
+        }
+
+        /// <inheritdoc />
+        public virtual string NormalizeValue(string value)
+        {
+            return value;
         }