+++ ContentPatcher/ModEntry.cs
@@ -51,7 +51,8 @@ namespace ContentPatcher
             new Migration_1_15_Rewrites(content),
             new Migration_1_16(),
             new Migration_1_17(),
-            new Migration_1_18()
+            new Migration_1_18(),
+            new Migration_1_19()
         };
 
         /// <summary>The special validation logic to apply to assets affected by patches.</summary>
@@ -176,6 +177,15 @@ namespace ContentPatcher
             this.UpdateContext(ContextUpdateType.All);
         }
 
+        /// <summary>The method invoked when the in-game clock changes.</summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event data.</param>
+        private void OnTimeChanged(object sender, TimeChangedEventArgs e)
+        {
+            this.Monitor.VerboseLog("Updating context: clock changed.");
+            this.UpdateContext(ContextUpdateType.OnTimeChange);
+        }
+
         /// <summary>The method invoked when the player warps.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event data.</param>
@@ -264,6 +274,7 @@ namespace ContentPatcher
                 helper.Events.Input.ButtonPressed += this.OnButtonPressed;
             helper.Events.GameLoop.ReturnedToTitle += this.OnReturnedToTitle;
             helper.Events.GameLoop.DayStarted += this.OnDayStarted;
+            helper.Events.GameLoop.TimeChanged += this.OnTimeChanged;
             helper.Events.Player.Warped += this.OnWarped;
             helper.Events.Specialized.LoadStageChanged += this.OnLoadStageChanged;